+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add the version app in the footer of the project.
 - Add `relative-time` component to table of inventory.
 - Add label to the table of inventory the last update.
+- Add download button to the image of the product.
 
 ### Changed
 - Change the color of the footer in mobile.